Remove unnecesary useSnackbar hook
